+++ src/core/files.py
@@ -83,7 +83,7 @@ default_general = {
     "allowable_azure_downtime_minutes": 60,
     "send_message_to_local_file_only": False,
     "e2_tcp_delay_milliseconds": 1,
-    "fail_backoff_seconds": 3600,
+    "lock_reset_seconds": 43200,
     "e2_buffer_length": 75,
 }