+++ src/main.py
@@ -3,12 +3,17 @@ import logging
 from store import store
 import os
 from datetime import datetime
+import json
+import core
+import time
 
-__version__: str = "v0.14.0"
+__version__: str = "v0.15.0"
 __author__: str = "Henry West <west.henrymiles@gmail.com>"
 
 logger = logging.getLogger(__name__)
 
+with open(core.GENERAL_SETTINGS, "r") as f:
+    general_settings = json.load(f)
 
 def main() -> None:
     """
@@ -24,6 +29,23 @@ def main() -> None:
             )
         except Exception as e:
             logger.error(f"Cannot write version file: {e}")
+
+    try:
+        with open("lock.json", "r") as f:
+            contents = json.load(f)
+            timeobj = datetime.fromisoformat(contents["timestamp"])
+            current_time = datetime.now()
+            time_diff = current_time - timeobj
+            seconds_elapsed = time_diff.total_seconds()
+            if seconds_elapsed < general_settings.get("lock_reset_seconds", 43_200):
+                logger.critical(f"lock file exists with a young timestamp. shutting down the driver immediately.")
+                time.sleep(1)
+                os._exit(1)
+    except FileNotFoundError:
+        logger.warning(f"lock file does not exist. it is probably okay to continue...")
+    except Exception as e:
+        logger.error(f"Could not check lock file: {e}")
+
 
     try:
         asyncio.run(store.mainloop())